move TypeBasedSPI to type package
