+++ encrypt-core/encrypt-core-common/src/main/java/org/apache/shardingsphere/encrypt/strategy/spi/Encryptor.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.encrypt.strategy.spi;
 
-import org.apache.shardingsphere.spi.TypeBasedSPI;
+import org.apache.shardingsphere.spi.type.TypeBasedSPI;
 
 /**
  * Encryptor.
