+++ shardingsphere-spi/src/main/java/org/apache/shardingsphere/spi/type/TypeBasedSPI.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.spi;
+package org.apache.shardingsphere.spi.type;
 
 import java.util.Properties;
