+++ shardingsphere-spi/src/test/java/org/apache/shardingsphere/spi/fixture/TypeBasedSPIFixture.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.spi.fixture;
 
-import org.apache.shardingsphere.spi.TypeBasedSPI;
+import org.apache.shardingsphere.spi.type.TypeBasedSPI;
 
 public interface TypeBasedSPIFixture extends TypeBasedSPI {
 }
