+++ shardingsphere-spi/src/test/java/org/apache/shardingsphere/spi/fixture/TypeBasedSPIFixtureServiceLoader.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.spi.fixture;
 
 import org.apache.shardingsphere.spi.NewInstanceServiceLoader;
-import org.apache.shardingsphere.spi.TypeBasedSPIServiceLoader;
+import org.apache.shardingsphere.spi.type.TypeBasedSPIServiceLoader;
 
 public final class TypeBasedSPIFixtureServiceLoader extends TypeBasedSPIServiceLoader<TypeBasedSPIFixture> {
